+++ src/javascript.ts
@@ -1,7 +1,6 @@
 import {javascript as js, javascriptLanguage, scopeCompletionSource} from '@codemirror/lang-javascript';
 import type {Extension} from '@codemirror/state';
 
-export default (): Extension => [
-	js(),
-	javascriptLanguage.data.of({autocomplete: scopeCompletionSource(globalThis)}),
-];
+export const jsCompletion = javascriptLanguage.data.of({autocomplete: scopeCompletionSource(globalThis)});
+
+export default (): Extension => [js(), jsCompletion];